Started writing 'failure' tests for extension terminals.
Finally I have realized that the current approach to partitioning and
checking syntax is deficient. It is not right to first roughly split
clauses into classes by guessing their true class, and then verify our
guess.

Backtracking will be hard to implement with syntax-rules, and we
should backtrack if our guess was wrong but the clause may be still
syntactically correct. A better approach will be to first filter our
entirely correct clauses, then check for remaining ones, guess their
classes, and then produce some meaningful error message.

Dunno about multiple errors. `syntax-error` does not give a chance
to do it, so I guess it will be just the first erroneous clause to be
reported.