Reimplementated extension terminal partitioning in a new way.
All failure tests actually pass. Success tests do not, because
valid extension terminals are not properly postprocessed yet.
This will be done together with some refactoring of the code.

Also, tests tweaks because I have found more peculiar clauses.

Also x2, tests are incredibly slooooow now. Need to do something
about it in the future. I do not want the define-language to be
as slow as these tests are.