Broken down test suite into pieces
These are running too long. Now success and failure tests will be
executed for subdirectories to make intermediate results visible.

This change demands that new tests are to be explicitly added to
the '.travis.yml' file for them to be run by the CI.

Possibly should consider breaking tests into even more pieces...