Implemented filtering of standalone nonterminals
All standalone tests actually pass, both success and failure ones.
Extension test expansion fails the build. Simply ignore it.

It was somewhat challenging to get the patterns for all standalone
cases in the right order. We have only one chance to match, after
that we can only verify our guess. I still have a feeling that this
is not an ideal solution, but it seems to work and cover all cases.

Also fixed the test runner to do not break on ':' in error messages.