Implemented toplevel clause partitioning
Rewritten it in a new style with %verify functions. It has a pretty
peculiar control flow in 'high-level' part (partitioning-toplevel.sld).
Well, maybe that's because of the checks for uniqueness and so on.

I believe, that boilerplate generator from structure-toplevel.sld can
be moved into utilities. At least to generate $is-a, $must-be, $verify
functions; IIRC, they are all the same everywhere.

Also fixed a typo in tests.