Updated standalone nonterminal partitioning
- Started unifying names
- Fixed naming in 'business logic' layer
- Updated checker and verifier implementations
- Fixed some incorrect error messages in tests

Also fixed a peculiar issue with matching in common utilities.
The thing is, (a ... . d) actually matches _everything_, at least
with Chibi. I am even in doubt whether this is a legal pattern.

It _should_ match this pattern:

  (<pattern> ... <pattern> <ellipsis> <pattern> ... . <pattern>)

like this:

  (                  a         ...                  .     d    )

which is equivalent to

  (   P1 ... Pk      Pe    <ellpsis>   Pm+1 ... Pn  .    Px    )

with k, m, n all equal zero and thus Px matching anything.

I guess the reason is in the official definition of 'improper list'
as 'anything that is not a proper list' where a 'proper list' is
either an empty list, or a pair which cdr is a proper list. Thus,
_technically_ anything can be treated as an improper list, without
it being a pair. So a proper way to match improper list is this:
(x y ... . d).