Implemented 'define-language' interface syntax
Nuff said.

The error message for invalid lanugage definitions had to be
changed from "Invalid language definition: (define-language)"
to just "Invalid language definition" because it is not possible
for a macro to refer to its own form.

Technically, we can do (syntax-error "Invalid language definition"
(define-language . expressions)), but this will output a syntactic
closure of 'define-language', not the form that was in the user's
program.

So it has been replaced with a more simple error message. I don't
think this will be too cryptic do understand (given that a line
number is still provided to the user).