Simplify nonterminal predicate syntax
It does not add any syntactical value for the user, so I have
removed it. Instead we use this better syntax for nonterminals:

  (Name (maybe Predicate) (meta-vars ...) productions ...)

That is, if the second element is a list, it is non-predicated
form. If the third element is a list, then the second must be
a predicate name. Everything else is just incorrect nonterminal.

Yes, "expected something to be a list" and "unexpected dotted pair"
are going to be removed. They do not make sense, the user must
deliberately write invalid syntax to see those error messages.

Currently, some test cases fail at this commit. This will be fixed
in the following ones.