Stricter list syntax requirements
This removes all obscure "expected something to be a proper list"
and "unexpected dotted list" errors. The reason is that such messages
are hardly informative, and to get them the user must deliberately
write dotted lists in unexpected places. The only place where they
are allowed are nonterminal production definitions.

Instead we display a generic "invalid syntax" message. I do not think
this is going to be more obscure.

Note that here we leave dotted lists in $can-be? predicates which
match literal clauses. This is done to prevent the partitioning
code from treating invalid 'easily recognizable' forms as something
different, which often results into really obscure error messages.