Syntax error testing utilities
ENOUGH OF THIS SHIT! TWENTY MINUTES TO RUN THE FUCKING UNIT-TESTS
IS A RIDICULOUS AMOUNT OF TIME. FUCK THE PORTABILITY. DON'T NEED IT!

Sorry, but it *really* irritates me. Or, I'd rather say 'irritated'.

This adds `assert-syntax-error` assertion for (te) testing library.
It enables doing what tools/run-failure-tests does, but avoids
reloading and restarting Chibi for every single failure test which
hopefully will dramatically decrease time needed to run them all.

The form is used like this:

    (define-syntax foo
      (syntax-rules ()
        ((_) (syntax-error "Example" 42)) ) )

    (define-test-case (my-test-case)
      (define-test ()
        (assert-syntax-error ("Example" 42)
          (foo) ) ) )

It is Chibi-specific, but has a clean interface, so it could be
made portable (provided that implementations allow this kind of
stunt with macroexpansion).