Typechecking
This implements initial type validity checking so that further
processing procedures could safely assume it.

The utility macro is purposely simplistic. It _could_ be
extended to handle multi-argument predicates, but this would
greatly increase code complexity (we will need to generate
and use an arbitrary number of temporary binding names).

Also, this prominently shows how verbose record accessors are.
I thought about how to deal with it, but with Scheme there seems
to be no good way to do this. So, I believe, I will leave accessor
names as is and then add a destructuring macro like `with-terminal`
to bind all fields at once (if I ever need to do this).