Generalize type checking
My previous statement was a lie. *This* is how I want generic error
checking to be. There should be no pure checkers or something like
that. Every procedure will do meaningful job. If it encounters an
error, it raises a continuable exception to report it, substitutes
some reasonable default value, and goes on with whatever it does.

Errors are then collected with an exception handler similar to the
one used in tests, and reported to the user. If we saw an error, we
know that the processing results are unreliable.

Therefore, processing procedures must be ready for errors, and for
default values. I do not really know how to handle some errors in
a proper way. E.g., invalid names: now I use just <invalid-name>
symbol for this, but I will need to do uniqueness test on the names
and it would be stupid to report these ones. Maybe it will be exported.

Another point are continuable exceptions. It could have been coded
to use lambda callback handlers, but this would mean threading this
callback context through the whole code. Dynamic variables are another
option, but I believe this is not the use-case they are intended for.

Also, some points of the code (like name? or meta-var?) may seem to
be code duplicating, but that's the whole point behind tests: one
effectively writes the same thing twice using different languages,
and this can break your build if it is not changed consistently.

And finally, `check` macro is highly local, made just for better
readability of type checking rules, defaults, and errors. Somewhat
complex, but it does its job.