Meta-variable suffix trimmer
This procedure is needed for production processing to trim optional
meta-variable suffixes as described by Nanopass spec: a meta-variable
can be suffixed by an arbitrary number of decimal digits followed by
an arbitrary number of special characters (?+*^).

This implies that actual meta-variable names do not end with such
suffixes. E.g., if one defines meta-variables 'e' and 'e1' then 'e1'
or 'e1*' could refer to any of them. We will detect and prohibit
such names later.

The implementation uses blatantly inefficient looping over a string
via indexing, but Chibi does not have SRFI-13 support and I do not
want to depend on Chibi's unportable and unfinished regexp engine.

Anyway, this procedure would not be on the hot path (like, called
a handful of times per symbol of a language), and typical input for
it would be around 1--5 ASCII characters. Efficiency doesn't seem to
really matter here. There is regexp for anyone willing to optimize.