Drop support of short terminal form
I mean that one where terminal name is inferred from the predicate
which is required to be specified as variable binding:

    (define-language Lang
      (terminals
        (number? (n)) ) )

Initially I thought this would be a good shortcut for terminals like
this, which are based on some pre-existing predicates.

However, this led to another idea of treating 'number' and 'number?'
as equal with intent to avoid referring to names with question mark.
But such special comparison itself leads to issues. And I still think
that terminals named 'number?' are bad.

So I have decided to ditch this extension away. The user now must
specify terminal names explicitly. This avoids the issue, allows to
compare names with eq?, and shaves off some macro code.